meta-xilinx-core: qemu-xilinx-multiarch-helper-native fix default help
Code change is based on the discussion with Corey Thompson on the
meta-xilinx@lists.yoctoproject.org mailing list.

It was pointed out that the default help behavior did not match what runqemu
was expecting from the actual qemu execution.  We need to properly handle
the user asking from help information or improperly invoking the binary with
correct return codes.

The new behavior will ensure that incorrect arguments results in a non-zero
return code with an error message or basic help text.

If the user passes -h, -help or --help (matching qemu's behavior) will
report the basic help text, and then invoke the APU, PMU or PLM qemu as
appropriate to augment and provide the full set of arguments.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>